feat(perf-views) Add release markers to duration breakdown
Add release markers to the release breakdown chart. This will help with
diagnosing performance changes related to releases.

I've moved decodeScalar() out of the eventsv2 views directory as it is
used in other view paths as well.
